+++ src/main/java/com/tac/guns/client/render/gun/model/ak47_animation.java
@@ -41,8 +41,8 @@ public class ak47_animation extends SkinAnimationModel {
         GunSkin skin = SkinManager.getSkin(stack);
 
         Gun gun = ((GunItem) stack.getItem()).getGun();
-        int gunRate = (int) Math.min(ShootingHandler.calcShootTickGap(gun.getGeneral().getRate()), 4);
-        int rateBias = (int) (ShootingHandler.calcShootTickGap(gun.getGeneral().getRate()) - gunRate);
+        float gunRate = Math.min(ShootingHandler.calcShootTickGap(gun.getGeneral().getRate()), 4);
+        float rateBias = ShootingHandler.calcShootTickGap(gun.getGeneral().getRate()) - gunRate;
         float cooldownOg = (ShootingHandler.get().getshootMsGap() - rateBias) / gunRate < 0 ? 1 : MathHelper.clamp((ShootingHandler.get().getshootMsGap() - rateBias) / gunRate, 0, 1);
 
         matrices.push();
@@ -67,8 +67,8 @@ public class ak47_animation extends SkinAnimationModel {
 
             /*//We're getting the cooldown tracker for the item - items like the sword, ender pearl, and chorus fruit all have this too.
             Gun gun = ((GunItem) stack.getItem()).getGun();
-        int gunRate = (int) Math.min(ShootingHandler.calcShootTickGap(gun.getGeneral().getRate()), 4);
-int rateBias = (int) (ShootingHandler.calcShootTickGap(gun.getGeneral().getRate()) - gunRate);
+        float gunRate = Math.min(ShootingHandler.calcShootTickGap(gun.getGeneral().getRate()), 4);
+float rateBias = ShootingHandler.calcShootTickGap(gun.getGeneral().getRate()) - gunRate;
         float cooldownOg = (ShootingHandler.get().getshootMsGap() - rateBias) / gunRate < 0 ? 1 : MathHelper.clamp((ShootingHandler.get().getshootMsGap() - rateBias) / gunRate, 0, 1);
         */
             if (transformType.isFirstPerson()) {
